+++ examples/postly/lib/domain/repository/post_repository.dart
@@ -1,0 +1,7 @@
+import 'package:dartz/dartz.dart';
+import '../entity/post.dart';
+
+abstract class PostRepository {
+  Future<Either<Exception, List<Post>>> getPosts();
+  Future<Either<Exception, Post>> getPost(int id);
+}