+++ cherrypick/CHANGELOG.md
@@ -1,5 +1,17 @@
+## 2.1.0-dev.1
+
+ - **FIX**: fix warnings.
+ - **FIX**: fix warnings.
+ - **FIX**: support passing params when resolving dependency recursively in parent scope.
+ - **FEAT**: Add async dependency resolution and enhance example.
+ - **FEAT**: Add async dependency resolution and enhance example.
+
 ## 2.0.2
 - **FIX**: support passing params when resolving dependency recursively in parent scope.
+
+## 2.1.0-dev.0
+
+ - **FEAT**: Add async dependency resolution and enhance example.
 
 ## 2.0.1
 - **FIX**: fix warning.